Introduce Allocator and VecLike traits over BufferPool
Move the BufferPool/PoolVec implementation into a new buffer_pool submodule
of binius-compute, then add two traits at the crate root that abstract them:

- VecLike<T>: the growable, Vec-like buffer surface the prover relies on
  (capacity, push, clear, resize, extend_from_slice, spare_capacity_mut,
  set_len) plus Deref/DerefMut to [T] and Extend<T>. Implemented for PoolVec.
- Allocator: a source of VecLike buffers, with an associated GAT
  Vec<T>: VecLike<T> and alloc<T>(&self, capacity) -> Self::Vec<T>. The
  alloc method carries no Pod bound. Implemented for &'alloc BufferPool with
  Vec<T> = PoolVec<'alloc, T>.

The Pod bound is dropped from alloc_vec/PoolVec throughout; growth methods
that need it now carry a T: Clone bound (matching Vec), and PoolVec::drop
runs element destructors via clear() before reclaiming the block so a
non-Pod, droppable T no longer leaks (a no-op for Copy types like Word).
bytemuck is no longer a dependency of binius-compute.

Refactor the prover's allocation code to be generic over &A where
A: Allocator: IOPProver::prove takes alloc: &A, the And/Mul/BinMul witness
structs and their build_/prove_ functions are parameterized by A with
buffers typed A::Vec<Word>, and both call sites pass &(&pool) so that
A = &BufferPool.
