+++ crates/compute/Cargo.toml
@@ -8,4 +8,3 @@ authors.workspace = true
 workspace = true
 
 [dependencies]
-bytemuck = { workspace = true }
