+++ crates/prover/src/prove.rs
@@ -3,7 +3,7 @@
 
 use std::marker::PhantomData;
 
-use binius_compute::{BufferPool, PoolVec};
+use binius_compute::{Allocator, BufferPool, VecLike};
 use binius_core::{
 	constraint_system::{
 		AndConstraint, BmulConstraint, ConstraintSystem, ImulConstraint, ValueVec,
@@ -88,13 +88,14 @@ impl IOPProver {
 	///
 	/// This is the core proving logic, independent of the specific IOP compilation strategy.
 	/// For most users, [`Prover::prove`] is the simpler interface.
-	pub fn prove<P, Channel>(
+	pub fn prove<A, P, Channel>(
 		&self,
 		witness: ValueVec,
 		channel: &mut Channel,
-		pool: &BufferPool,
+		alloc: &A,
 	) -> Result<(), Error>
 	where
+		A: Allocator,
 		P: PackedField<Scalar = B128>,
 		Channel: IOPProverChannel<P>,
 	{
@@ -135,9 +136,9 @@ impl IOPProver {
 			)
 			.entered();
 			let mul_witness = tracing::debug_span!("Assemble columns")
-				.in_scope(|| build_intmul_witness(&cs.imul_constraints, &witness, pool));
+				.in_scope(|| build_intmul_witness(&cs.imul_constraints, &witness, alloc));
 
-			let intmul_output = prove_intmul_reduction::<_, P, _>(mul_witness, &mut *channel)?;
+			let intmul_output = prove_intmul_reduction::<_, _, P, _>(mul_witness, &mut *channel)?;
 			drop(intmul_guard);
 			Some(intmul_output)
 		} else {
@@ -157,9 +158,9 @@ impl IOPProver {
 			)
 			.entered();
 			let binmul_witness = tracing::debug_span!("Assemble columns")
-				.in_scope(|| build_binmul_witness(&cs.bmul_constraints, &witness, pool));
+				.in_scope(|| build_binmul_witness(&cs.bmul_constraints, &witness, alloc));
 
-			let binmul_output = prove_binmul_reduction::<_, P, _>(binmul_witness, &mut *channel);
+			let binmul_output = prove_binmul_reduction::<_, _, P, _>(binmul_witness, &mut *channel);
 			drop(binmul_guard);
 			Some(binmul_output)
 		} else {
@@ -172,7 +173,7 @@ impl IOPProver {
 				.entered();
 		let bitand_claim = {
 			let bitand_witness = tracing::debug_span!("Assemble columns")
-				.in_scope(|| build_bitand_witness(&cs.and_constraints, &witness, pool));
+				.in_scope(|| build_bitand_witness(&cs.and_constraints, &witness, alloc));
 
 			let AndCheckOutput {
 				a_eval,
@@ -180,7 +181,7 @@ impl IOPProver {
 				c_eval,
 				z_challenge,
 				eval_point,
-			} = prove_bitand_reduction::<B128, P, _>(bitand_witness, &mut *channel);
+			} = prove_bitand_reduction::<_, B128, P, _>(bitand_witness, &mut *channel);
 			OperatorData {
 				evals: vec![a_eval, b_eval, c_eval],
 				r_zhat_prime: z_challenge,
@@ -407,10 +408,12 @@ where
 		let mut channel = self
 			.basefold_compiler
 			.create_channel_without_zk_from_transcript::<H, Challenger_, _>(transcript);
-		// Working buffers for this proof are drawn from a single pool that lives for the call.
+		// Working buffers for this proof are drawn from a single pool that lives for the call. The
+		// pool is passed as an `&BufferPool` allocator.
 		let pool = BufferPool::new();
+		let alloc = &pool;
 		self.iop_prover
-			.prove::<P, _>(witness, &mut channel, &pool)?;
+			.prove::<_, P, _>(witness, &mut channel, &alloc)?;
 		channel.finish();
 		Ok(())
 	}
@@ -484,11 +487,12 @@ pub fn pack_witness<P: PackedField<Scalar = B128>>(
 	Ok(FieldBuffer::new(log_witness_elems, padded_witness_elems))
 }
 
-fn prove_bitand_reduction<F, PChallenge, Channel>(
-	witness: AndCheckWitness<'_>,
+fn prove_bitand_reduction<A, F, PChallenge, Channel>(
+	witness: AndCheckWitness<A>,
 	channel: &mut Channel,
 ) -> AndCheckOutput<F>
 where
+	A: Allocator,
 	F: BinaryField + From<B8>,
 	PChallenge: PackedField<Scalar = F>,
 	Channel: binius_ip_prover::channel::IPProverChannel<F>,
@@ -514,11 +518,12 @@ where
 	prover.prove_with_channel(channel)
 }
 
-fn prove_intmul_reduction<F, P, Channel>(
-	witness: MulCheckWitness<'_>,
+fn prove_intmul_reduction<A, F, P, Channel>(
+	witness: MulCheckWitness<A>,
 	channel: &mut Channel,
 ) -> Result<IntMulOutput<F>, Error>
 where
+	A: Allocator,
 	F: BinaryField<Underlier: Divisible<u64>>,
 	P: PackedField<Scalar = F>,
 	Channel: IOPProverChannel<P>,
@@ -533,11 +538,12 @@ where
 	Ok(mulcheck_prover.prove(intmul_witness))
 }
 
-fn prove_binmul_reduction<F, P, Channel>(
-	witness: BinMulCheckWitness<'_>,
+fn prove_binmul_reduction<A, F, P, Channel>(
+	witness: BinMulCheckWitness<A>,
 	channel: &mut Channel,
 ) -> BinMulOutput<F>
 where
+	A: Allocator,
 	F: BinaryField + From<u128>,
 	P: PackedField<Scalar = F>,
 	Channel: binius_ip_prover::channel::IPProverChannel<F>,
@@ -565,38 +571,38 @@ where
 	prove_binmul::<F, P, _>(&binmul_witness, channel)
 }
 
-struct AndCheckWitness<'alloc> {
-	a: PoolVec<'alloc, Word>,
-	b: PoolVec<'alloc, Word>,
-	c: PoolVec<'alloc, Word>,
+struct AndCheckWitness<A: Allocator> {
+	a: A::Vec<Word>,
+	b: A::Vec<Word>,
+	c: A::Vec<Word>,
 }
 
-struct MulCheckWitness<'alloc> {
-	a: PoolVec<'alloc, Word>,
-	b: PoolVec<'alloc, Word>,
-	lo: PoolVec<'alloc, Word>,
-	hi: PoolVec<'alloc, Word>,
+struct MulCheckWitness<A: Allocator> {
+	a: A::Vec<Word>,
+	b: A::Vec<Word>,
+	lo: A::Vec<Word>,
+	hi: A::Vec<Word>,
 }
 
-struct BinMulCheckWitness<'alloc> {
-	a_lo: PoolVec<'alloc, Word>,
-	a_hi: PoolVec<'alloc, Word>,
-	b_lo: PoolVec<'alloc, Word>,
-	b_hi: PoolVec<'alloc, Word>,
-	c_lo: PoolVec<'alloc, Word>,
-	c_hi: PoolVec<'alloc, Word>,
+struct BinMulCheckWitness<A: Allocator> {
+	a_lo: A::Vec<Word>,
+	a_hi: A::Vec<Word>,
+	b_lo: A::Vec<Word>,
+	b_hi: A::Vec<Word>,
+	c_lo: A::Vec<Word>,
+	c_hi: A::Vec<Word>,
 }
 
-fn build_bitand_witness<'alloc>(
+fn build_bitand_witness<A: Allocator>(
 	and_constraints: &[AndConstraint],
 	witness: &ValueVec,
-	pool: &'alloc BufferPool,
-) -> AndCheckWitness<'alloc> {
+	alloc: &A,
+) -> AndCheckWitness<A> {
 	let n_constraints = and_constraints.len();
 
-	let mut a = pool.alloc_vec::<Word>(n_constraints);
-	let mut b = pool.alloc_vec::<Word>(n_constraints);
-	let mut c = pool.alloc_vec::<Word>(n_constraints);
+	let mut a = alloc.alloc::<Word>(n_constraints);
+	let mut b = alloc.alloc::<Word>(n_constraints);
+	let mut c = alloc.alloc::<Word>(n_constraints);
 
 	(and_constraints, a.spare_capacity_mut(), b.spare_capacity_mut(), c.spare_capacity_mut())
 		.into_par_iter()
@@ -616,17 +622,17 @@ fn build_bitand_witness<'alloc>(
 	AndCheckWitness { a, b, c }
 }
 
-fn build_intmul_witness<'alloc>(
+fn build_intmul_witness<A: Allocator>(
 	imul_constraints: &[ImulConstraint],
 	witness: &ValueVec,
-	pool: &'alloc BufferPool,
-) -> MulCheckWitness<'alloc> {
+	alloc: &A,
+) -> MulCheckWitness<A> {
 	let n_constraints = imul_constraints.len();
 
-	let mut a = pool.alloc_vec::<Word>(n_constraints);
-	let mut b = pool.alloc_vec::<Word>(n_constraints);
-	let mut lo = pool.alloc_vec::<Word>(n_constraints);
-	let mut hi = pool.alloc_vec::<Word>(n_constraints);
+	let mut a = alloc.alloc::<Word>(n_constraints);
+	let mut b = alloc.alloc::<Word>(n_constraints);
+	let mut lo = alloc.alloc::<Word>(n_constraints);
+	let mut hi = alloc.alloc::<Word>(n_constraints);
 
 	(
 		imul_constraints,
@@ -654,19 +660,19 @@ fn build_intmul_witness<'alloc>(
 	MulCheckWitness { a, b, lo, hi }
 }
 
-fn build_binmul_witness<'alloc>(
+fn build_binmul_witness<A: Allocator>(
 	bmul_constraints: &[BmulConstraint],
 	witness: &ValueVec,
-	pool: &'alloc BufferPool,
-) -> BinMulCheckWitness<'alloc> {
+	alloc: &A,
+) -> BinMulCheckWitness<A> {
 	let n_constraints = bmul_constraints.len();
 
-	let mut a_lo = pool.alloc_vec::<Word>(n_constraints);
-	let mut a_hi = pool.alloc_vec::<Word>(n_constraints);
-	let mut b_lo = pool.alloc_vec::<Word>(n_constraints);
-	let mut b_hi = pool.alloc_vec::<Word>(n_constraints);
-	let mut c_lo = pool.alloc_vec::<Word>(n_constraints);
-	let mut c_hi = pool.alloc_vec::<Word>(n_constraints);
+	let mut a_lo = alloc.alloc::<Word>(n_constraints);
+	let mut a_hi = alloc.alloc::<Word>(n_constraints);
+	let mut b_lo = alloc.alloc::<Word>(n_constraints);
+	let mut b_hi = alloc.alloc::<Word>(n_constraints);
+	let mut c_lo = alloc.alloc::<Word>(n_constraints);
+	let mut c_hi = alloc.alloc::<Word>(n_constraints);
 
 	(
 		bmul_constraints,
